+++ shared/src/iosMain/kotlin/com/murzify/bambuddyspool/core/network/BambuddyHttpClientFactory.ios.kt
@@ -1,0 +1,6 @@
+package com.murzify.bambuddyspool.core.network
+
+import io.ktor.client.engine.HttpClientEngineFactory
+import io.ktor.client.engine.darwin.Darwin
+
+actual fun platformBambuddyHttpEngineFactory(): HttpClientEngineFactory<*> = Darwin